apps/walletapi/apps.py
from __future__ import unicode_literals

from django.apps import AppConfig

from suit.apps import DjangoSuitConfig

class SuitConfig(DjangoSuitConfig):
    layout = 'vertical'

class WalletapiConfig(AppConfig):
    name = 'walletapi'
