requirements.txt
djangorestframework
markdown
django-filter
https://github.com/darklow/django-suit/tarball/v2
